+++ lib/ex_aws/auth.ex
@@ -3,6 +3,7 @@ defmodule ExAws.Auth do
 
   alias ExAws.Auth.Credentials
   alias ExAws.Auth.Signatures
+  alias ExAws.Request.Url
 
   @moduledoc false
 
@@ -139,7 +140,7 @@ defmodule ExAws.Auth do
   end
 
   defp signature(http_method, url, query, headers, body, service, datetime, config) do
-    path = url |> get_path() |> uri_encode()
+    path = url |> Url.get_path(service) |> Url.uri_encode()
     request = build_canonical_request(http_method, path, query, headers, body)
     string_to_sign = string_to_sign(request, service, datetime, config)
     Signatures.generate_signature_v4(service, config, datetime, string_to_sign)
