+++ lib/ex_aws/auth/utils.ex
@@ -1,35 +1,6 @@
 defmodule ExAws.Auth.Utils do
   @moduledoc false
 
-  def uri_encode(url) do
-    url
-    |> String.replace("+", " ")
-    |> URI.encode(&valid_path_char?/1)
-  end
-
-  def get_path(url) do
-    base =
-      url
-      |> URI.parse()
-      |> Map.put(:path, nil)
-      |> Map.put(:fragment, nil)
-      |> URI.to_string()
-      |> URI.parse()
-      |> URI.to_string()
-
-    url
-    |> String.split(base, parts: 2)
-    |> List.last()
-  end
-
-  # Space character
-  def valid_path_char?(?\ ), do: false
-  def valid_path_char?(?/), do: true
-
-  def valid_path_char?(c) do
-    URI.char_unescaped?(c) && !URI.char_reserved?(c)
-  end
-
   def hash_sha256(data) do
     :sha256
     |> :crypto.hash(data)
